Add an `append` method
This inserts HTML/XML content at the end *of the inside* of selected element(s).

Also, noted a behavioural quirk with PHP in that we cannot use a return-type of `this` when the class instance is of a parent class rather than the original class the method belongs to.

Added a `create_fragment` private method to share some behaviour between `setValue` and `append` that have to do similar tasks.